+++ rentcar/src/main/java/com/rentingapp/rentcar/modules/client/ClientService.java
@@ -1,6 +1,8 @@
 package com.rentingapp.rentcar.modules.client;
 
 
+import com.rentingapp.rentcar.modules.client.entity.Client;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
@@ -8,38 +10,45 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
+@RequiredArgsConstructor
 @Service
 public class ClientService {
-    @Autowired
-    ClientRepository clientRepository;
 
-    public Client addClient(Client client)
+    private final ClientRepository clientRepository;
+
+    Client addClient(Client client)
     {
         return clientRepository.save(client);
     }
-    public Iterable<Client> getClients()
+    Iterable<Client> getClients()
     {
         return clientRepository.findAll();
     }
-    public ResponseEntity<?> getClient(int id)
+    ResponseEntity<?> getClient(int id)
     {
         try{
-            Client client = clientRepository.findById(id).get();
-            return new ResponseEntity<Client>(client, HttpStatus.OK);
+            Client client = clientRepository.findById(id).orElseThrow();
+            return new ResponseEntity<>(client, HttpStatus.OK);
         }
         catch(NoSuchElementException e)
         {
-            return new ResponseEntity<String>("Client with ID: " + id + " doesn't exist",HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Client with ID: " + id + " doesn't exist",HttpStatus.NOT_FOUND);
         }
     }
 
-    public ResponseEntity<?> deleteClient(int id){
+    ResponseEntity<String> deleteClient(int id){
+
         try{
+            Client client = clientRepository.findById(id).orElseThrow();
+            if(client.isRenting())
+            {
+                return new ResponseEntity<>("This client is renting a car thus he cannot be deleted",HttpStatus.OK);
+            }
             clientRepository.deleteById(id);
-            return new ResponseEntity<String>("Client with ID: " + id + " was deleted", HttpStatus.OK);
+            return new ResponseEntity<>("Client with ID: " + id + " was deleted", HttpStatus.OK);
         } catch (EmptyResultDataAccessException e)
         {
-            return new ResponseEntity<String>("Client with ID: " + id + " doesn't exist", HttpStatus.OK);
+            return new ResponseEntity<>("Client with ID: " + id + " doesn't exist", HttpStatus.OK);
         }
     }
 }
